% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_v2.R
\name{cochran_q_diagnostic}
\alias{cochran_q_diagnostic}
\title{Cochran Q heterogeneity diagnostic}
\usage{
cochran_q_diagnostic(beta_exp, beta_out, se_out, gamma, R = NULL)
}
\description{
Cochran Q heterogeneity diagnostic
}
