% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{compute_trait_covariance}
\alias{compute_trait_covariance}
\title{Compute Trait Covariance Matrix}
\usage{
compute_trait_covariance(n_trait, rho = 0.5)
}
\arguments{
\item{n_trait}{Number of traits}

\item{rho}{Correlation parameter}
}
\value{
Covariance matrix (n_trait x n_trait)
}
\description{
Creates an AR(1) structured covariance matrix.
}
\keyword{internal}
