% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{convert_geno_to_qtlnetwork}
\alias{convert_geno_to_qtlnetwork}
\title{Convert genotype data to QTLNetwork format}
\usage{
convert_geno_to_qtlnetwork(G, snp_info, output_prefix)
}
\arguments{
\item{G}{Genotype matrix (n_ind x n_snp), standardized}

\item{snp_info}{SNP information data frame}

\item{output_prefix}{Output file prefix}
}
\value{
NULL (writes files)
}
\description{
Convert genotype data to QTLNetwork format
}
