% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{convert_pheno_to_qtlnetwork}
\alias{convert_pheno_to_qtlnetwork}
\title{Convert phenotype data to QTLNetwork format}
\usage{
convert_pheno_to_qtlnetwork(pheno_data, trait_name, output_prefix)
}
\arguments{
\item{pheno_data}{Phenotype data frame with GID and trait columns}

\item{trait_name}{Trait name to process}

\item{output_prefix}{Output file prefix}
}
\value{
NULL (writes files)
}
\description{
Convert phenotype data to QTLNetwork format
}
