% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{decompose_true_effects}
\alias{decompose_true_effects}
\title{Decompose true effects into independent and shared components}
\usage{
decompose_true_effects(beta, V)
}
\arguments{
\item{beta}{Effect matrix (n_snp x n_trait)}

\item{V}{Trait covariance matrix (n_trait x n_trait)}
}
\value{
List with beta_ind and beta_shared matrices
}
\description{
Decompose true effects into independent and shared components
}
\keyword{internal}
