% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{default_effect_config}
\alias{default_effect_config}
\title{Default Effect Configuration}
\usage{
default_effect_config(n_trait, n_snp)
}
\arguments{
\item{n_trait}{Number of traits}

\item{n_snp}{Number of SNPs}
}
\value{
List with default effect configuration
}
\description{
Creates a default genetic architecture for testing.
}
\keyword{internal}
