% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-condped-data.R
\name{.block_cor}
\alias{.block_cor}
\title{Block correlation matrix (within-block 0.6, between-block 0.2)}
\usage{
.block_cor(m, within = 0.6, between = 0.2)
}
\description{
Block correlation matrix (within-block 0.6, between-block 0.2)
}
\keyword{internal}
