% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-condped-data.R
\name{.check_correlation_matrix}
\alias{.check_correlation_matrix}
\title{Validate a correlation matrix argument}
\usage{
.check_correlation_matrix(R, m, name)
}
\description{
Validate a correlation matrix argument
}
\keyword{internal}
