% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan-mt-omnibus.R
\name{.combine_effects}
\alias{.combine_effects}
\title{Combine per-marker effect lists into scan/estimate return format}
\usage{
.combine_effects(effects_list)
}
\arguments{
\item{effects_list}{List of outputs from \code{\link[=.format_one_effect]{.format_one_effect()}}.}
}
\value{
List with \code{effects_long}, \code{beta}, \code{covariance}.
}
\description{
Combine per-marker effect lists into scan/estimate return format
}
\keyword{internal}
