% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-condped-data.R
\name{.cov_from_h2}
\alias{.cov_from_h2}
\title{Build a covariance matrix from marginal variances and a correlation matrix}
\usage{
.cov_from_h2(var, R)
}
\description{
Build a covariance matrix from marginal variances and a correlation matrix
}
\keyword{internal}
