% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-condped-data.R
\name{.exchangeable_cor}
\alias{.exchangeable_cor}
\title{Exchangeable correlation matrix}
\usage{
.exchangeable_cor(m, r)
}
\arguments{
\item{m}{Dimension.}

\item{r}{Common correlation.}
}
\value{
An \verb{m x m} correlation matrix.
}
\description{
Exchangeable correlation matrix
}
\keyword{internal}
