% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-condped-simulation.R
\name{.family_sets_at}
\alias{.family_sets_at}
\title{Family sets (list of trait vectors, named by key) at one tolerance}
\usage{
.family_sets_at(tab, tol)
}
\description{
Family sets (list of trait vectors, named by key) at one tolerance
}
\keyword{internal}
