% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-condped-simulation.R
\name{.loci_from_signals}
\alias{.loci_from_signals}
\title{Derive a locus table from a normalised signal table}
\usage{
.loci_from_signals(sig_tab)
}
\arguments{
\item{sig_tab}{Normalised signal table.}
}
\description{
Derive a locus table from a normalised signal table
}
\keyword{internal}
