% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-associated-loci.R
\name{.merge_interval_list}
\alias{.merge_interval_list}
\title{Merge overlapping same-chromosome intervals}
\usage{
.merge_interval_list(intervals)
}
\arguments{
\item{intervals}{List of interval lists with \code{chromosome}, \code{lo},
\code{hi}, \code{selected}.}
}
\value{
Merged list (selected markers unioned).
}
\description{
Merge overlapping same-chromosome intervals
}
\keyword{internal}
