% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-condped-data.R
\name{.min_eigen_sym}
\alias{.min_eigen_sym}
\title{Smallest eigenvalue of a symmetric matrix}
\usage{
.min_eigen_sym(A)
}
\description{
Smallest eigenvalue of a symmetric matrix
}
\keyword{internal}
