% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-condped-simulation.R
\name{.min_size_at}
\alias{.min_size_at}
\title{Minimum set size at one tolerance (Inf when empty)}
\usage{
.min_size_at(tab, tol)
}
\description{
Minimum set size at one tolerance (Inf when empty)
}
\keyword{internal}
