% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan-mt-omnibus.R
\name{.precompute_Ar}
\alias{.precompute_Ar}
\title{Precompute A_j \%*\% r_tilde_j for all rotated coordinates}
\usage{
.precompute_Ar(rot)
}
\arguments{
\item{rot}{The rotation object.}
}
\value{
An \verb{m x n} matrix whose column \code{j} is \code{A_j \%*\% r_tilde_j}.
}
\description{
Precompute A_j \%*\% r_tilde_j for all rotated coordinates
}
\keyword{internal}
