% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-condped-data.R
\name{.propose_B_Q}
\alias{.propose_B_Q}
\title{Propose the q x m effect matrix for one attempt}
\usage{
.propose_B_Q(spec, m, trait_names, Sigma_P, S_star, rho_star)
}
\description{
Propose the q x m effect matrix for one attempt
}
\keyword{internal}
