% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-mt-null.R
\name{.reml_starts}
\alias{.reml_starts}
\title{Build the list of log-Cholesky starting points}
\usage{
.reml_starts(Y, m, n_starts, h2_start, start_values, chol_floor)
}
\arguments{
\item{Y}{Phenotype matrix.}

\item{m}{Number of traits.}

\item{n_starts}{Number of starts (first is deterministic).}

\item{h2_start}{Heritability split for the first start.}

\item{start_values}{Optional explicit first start (theta vector or
\code{list(Sigma_G, Sigma_E)}).}

\item{chol_floor}{Diagonal floor used when projecting to PD.}
}
\value{
List of theta vectors.
}
\description{
Build the list of log-Cholesky starting points
}
\keyword{internal}
