% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-condped-simulation.R
\name{.save_replicate_atomic}
\alias{.save_replicate_atomic}
\title{Atomic save: write to a temporary file, then rename}
\usage{
.save_replicate_atomic(obj, file)
}
\description{
Atomic save: write to a temporary file, then rename
}
\keyword{internal}
