% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-condped-data.R
\name{.truth_tables_per_signal}
\alias{.truth_tables_per_signal}
\title{Truth tables for all signals of a replicate}
\usage{
.truth_tables_per_signal(
  B_Q,
  Sigma_P,
  tolerance,
  tolerances,
  locus_id,
  chr,
  pos,
  causal_ids,
  trait_names
)
}
\description{
Truth tables for all signals of a replicate
}
\keyword{internal}
