% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan-mt-omnibus.R
\name{.validate_null_fit}
\alias{.validate_null_fit}
\title{Validate a null fit object for S3}
\usage{
.validate_null_fit(null_fit)
}
\arguments{
\item{null_fit}{Candidate object.}
}
\description{
Validate a null fit object for S3
}
\keyword{internal}
