% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\docType{package}
\name{gwas-functions}
\alias{CondPED}
\alias{CondPED-package}
\alias{gwas-functions}
\title{GWAS Interface for CondPED using QTLNetwork}
\description{
This module provides integration between the conditional projection framework
and QTLNetwork for multi-trait GWAS analysis.
}
\details{
The workflow follows method.md:
\enumerate{
\item Marginal GWAS -> significant SNP set L
\item Forward conditional GWAS -> independent effects
\item Reverse conditional GWAS -> mediation test
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mzSUO/CondPED}
  \item Report bugs at \url{https://github.com/mzSUO/CondPED/issues}
}

}
\author{
\strong{Maintainer}: Mingzhe Suo \email{suomingzhe1@gmail.com}

}
