% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{parse_qtlnetwork_output}
\alias{parse_qtlnetwork_output}
\title{Parse QTLNetwork output}
\usage{
parse_qtlnetwork_output(pre_file, trait_name)
}
\arguments{
\item{pre_file}{Path to .pre output file}

\item{trait_name}{Trait name}
}
\value{
Data frame with QTL information
}
\description{
Parse QTLNetwork output
}
