% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{print.condped_simdata}
\alias{print.condped_simdata}
\title{Print Method for Simulated Data}
\usage{
print.condped_simdata(x, ...)
}
\arguments{
\item{x}{Object of class condped_simdata}

\item{...}{Additional arguments passed to print}
}
\description{
Print Method for Simulated Data
}
