% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{print_projection_summary}
\alias{print_projection_summary}
\title{Print summary of conditional projection result}
\usage{
print_projection_summary(cond_result)
}
\arguments{
\item{cond_result}{output from conditional_phenotype() or reverse_conditional()}
}
\value{
NULL (prints to console)
}
\description{
Print summary of conditional projection result
}
