% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlnetwork.R
\name{qtxnetwork.input.trans}
\alias{qtxnetwork.input.trans}
\title{qtxnetwork.input.trans
Convert data frames to QTXNetwork .gen + .phe format.}
\usage{
qtxnetwork.input.trans(
  geno_data,
  pheno_data,
  geno_output_prefix,
  pheno_output_prefix,
  population = c("RIL", "F2")
)
}
\arguments{
\item{geno_data}{Data frame: chr, snp_id, pos, Ind1, Ind2, ...
Genotype coding must match population type.}

\item{pheno_data}{Data frame: id, Trait1, Trait2, ...}

\item{geno_output_prefix}{Prefix for .gen file.}

\item{pheno_output_prefix}{Prefix for .phe file.}

\item{population}{"RIL" (default, no dominance) or "F2".}
}
\value{
NULL (invisible)
}
\description{
qtxnetwork.input.trans
Convert data frames to QTXNetwork .gen + .phe format.
}
