% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlnetwork.R
\name{qtxnetwork.layer1.screen}
\alias{qtxnetwork.layer1.screen}
\title{qtxnetwork.layer1.screen
Layer 1: classify SNPs by number of significant traits.}
\usage{
qtxnetwork.layer1.screen(qtl_output, alpha = 0.05)
}
\arguments{
\item{qtl_output}{Output from qtxnetwork.output.trans().}

\item{alpha}{Significance threshold. Default 0.05.}
}
\value{
List: class1, layer2, snp_count, full_table.
}
\description{
qtxnetwork.layer1.screen
Layer 1: classify SNPs by number of significant traits.
}
