% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlnetwork.R
\name{qtxnetwork.perform}
\alias{qtxnetwork.perform}
\title{qtxnetwork.perform
Execute QTXNetwork for 1D and/or 2D scanning.}
\usage{
qtxnetwork.perform(
  qtxnetwork_path,
  gen_file,
  phe_file,
  pre_file,
  scan_2d = FALSE,
  qtx_mode = 2
)
}
\arguments{
\item{qtxnetwork_path}{Path to QTXNetwork executable.}

\item{gen_file}{Input .gen file.}

\item{phe_file}{Input .phe file.}

\item{pre_file}{Output .pre file.}

\item{scan_2d}{Logical. TRUE enables 2D AA epistasis scanning. Default FALSE.}

\item{qtx_mode}{QTX mode. Default 2 (multi-trait).}
}
\value{
NULL (invisible)
}
\description{
qtxnetwork.perform
Execute QTXNetwork for 1D and/or 2D scanning.
}
