% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{reverse_conditional}
\alias{reverse_conditional}
\title{Construct reverse conditional phenotype}
\usage{
reverse_conditional(
  pheno_data,
  target_trait,
  exclude_trait,
  adjust_covariates = FALSE,
  covariate_cols = NULL
)
}
\arguments{
\item{pheno_data}{a data frame with GID and trait columns}

\item{target_trait}{character, name of the trait to be conditioned}

\item{exclude_trait}{character, name of the trait to exclude from conditioning set}

\item{adjust_covariates}{logical, whether to adjust for covariates (default: FALSE)}

\item{covariate_cols}{character vector, names of covariate columns}
}
\value{
same structure as conditional_phenotype()
}
\description{
Condition target trait on all traits EXCEPT the exclusion trait.
Used in reverse conditional GWAS to test mediation hypothesis.
}
\examples{
\dontrun{
# Test if QTL effect on Yield is mediated through Height
reverse_result <- reverse_conditional(pheno_data, 
                                     target_trait = "Yield",
                                     exclude_trait = "Height")
}
}
