% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{run_qtlnetwork}
\alias{run_qtlnetwork}
\title{Run QTLNetwork GWAS}
\usage{
run_qtlnetwork(
  qtxnetwork_path,
  gen_file,
  phe_file,
  output_prefix,
  verbose = TRUE
)
}
\arguments{
\item{qtxnetwork_path}{Path to QTLNetwork executable}

\item{gen_file}{Path to .gen file}

\item{phe_file}{Path to .phe file}

\item{output_prefix}{Output file prefix}

\item{verbose}{Print command (default: TRUE)}
}
\value{
NULL (writes output file)
}
\description{
Run QTLNetwork GWAS
}
