% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{sd_from_pve}
\alias{sd_from_pve}
\title{由 PVE 反推效应量标准差
RIL: Var(X)=4\emph{maf}(1-maf); F2: Var(X)=2\emph{maf}(1-maf)}
\usage{
sd_from_pve(pve, maf = 0.3, population = c("RIL", "F2"))
}
\description{
由 PVE 反推效应量标准差
RIL: Var(X)=4\emph{maf}(1-maf); F2: Var(X)=2\emph{maf}(1-maf)
}
