% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_genotype}
\alias{simulate_genotype}
\title{Generate Genotype Matrix}
\usage{
simulate_genotype(n_ind, n_snp, maf_range = c(0.05, 0.5), chr_snpNum = NULL)
}
\arguments{
\item{n_ind}{Number of individuals}

\item{n_snp}{Number of SNPs}

\item{maf_range}{Range for minor allele frequency}

\item{chr_snpNum}{Vector. Number of SNPs per chromosome (optional)}
}
\value{
List with G (genotype matrix) and snp_info (data frame)
}
\description{
Simulates genotype data with Hardy-Weinberg equilibrium.
Genotypes are coded as -1, 0, 1 (additive coding).
}
