% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_phenotype}
\alias{simulate_phenotype}
\title{Generate Phenotypes from Genotype and Effects}
\usage{
simulate_phenotype(G, beta, V, heritability, n_env = 1)
}
\arguments{
\item{G}{Genotype matrix (n_ind x n_snp)}

\item{beta}{True effect matrix (n_snp x n_trait)}

\item{V}{Trait covariance matrix}

\item{heritability}{Heritability value}

\item{n_env}{Number of environments}
}
\value{
List with Y (phenotype) and E (environmental effects)
}
\description{
Computes phenotypes as: Y = G * beta + E
where E ~ MVN(0, V * (1-h2)/h2 * sigma_g^2)
}
