% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{summary.condped_simdata}
\alias{summary.condped_simdata}
\title{Summary Method for Simulated Data}
\usage{
summary.condped_simdata(object, ...)
}
\arguments{
\item{object}{Object of class condped_simdata}

\item{...}{Additional arguments}
}
\description{
Summary Method for Simulated Data
}
