% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_v2.R
\name{weighted_median_mr}
\alias{weighted_median_mr}
\title{Weighted-median MR sensitivity estimate}
\usage{
weighted_median_mr(beta_exp, beta_out, se_out)
}
\description{
Weighted-median MR sensitivity estimate
}
