% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPED.R
\name{condPED}
\alias{condPED}
\title{condPED: Conditional Projection for Dissecting Pleiotropic Effects}
\usage{
condPED(
  pheno_data,
  geno_data,
  method = "qtlnetwork",
  output_dir = "./condPED_output",
  run_causal = TRUE,
  ...
)
}
\description{
唯一的用户入口函数，封装完整流程
}
