% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-utils.R
\name{.check_prob}
\alias{.check_prob}
\title{Check that an argument is a single probability in (0, 1)}
\usage{
.check_prob(x, name)
}
\arguments{
\item{x}{Value to check.}

\item{name}{Argument name for the error message.}
}
\description{
Check that an argument is a single probability in (0, 1)
}
\keyword{internal}
