% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-condped-data.R
\name{.sim_genotypes}
\alias{.sim_genotypes}
\title{Generate genotype dosages with optional group structure}
\usage{
.sim_genotypes(n, maf, structured, n_groups, fst)
}
\description{
Generate genotype dosages with optional group structure
}
\keyword{internal}
