% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-condped-data.R
\name{.symmetric_psd_sqrt}
\alias{.symmetric_psd_sqrt}
\title{Symmetric square root of a PSD matrix}
\usage{
.symmetric_psd_sqrt(Sigma)
}
\description{
Symmetric square root of a PSD matrix
}
\keyword{internal}
