% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{sim_genotype}
\alias{sim_genotype}
\title{模拟基因型矩阵}
\usage{
sim_genotype(n, p, maf = 0.3, population = c("RIL", "F2"), seed = NULL)
}
\arguments{
\item{n}{样本量}

\item{p}{SNP 数}

\item{maf}{次等位基因频率}

\item{population}{"RIL"(-1/1) 或 "F2"(-1/0/1)}
}
\description{
模拟基因型矩阵
}
