% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{sim_phenotype_M1}
\alias{sim_phenotype_M1}
\title{M1: 基础模型（无因果链）}
\usage{
sim_phenotype_M1(
  X,
  B,
  Sigma_E,
  epi_pairs = NULL,
  epi_effects = NULL,
  seed = NULL
)
}
\description{
M1: 基础模型（无因果链）
}
