% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{sim_phenotype_M2}
\alias{sim_phenotype_M2}
\title{M2: 因果链模型
Y = (X\emph{B + eps) \%}\% solve(I - t(Tau))
协方差传播: Sigma_Y = (I-Tau)^{-1} \%\emph{\% Sigma_base \%}\% (I-t(Tau))^{-1}}
\usage{
sim_phenotype_M2(
  X,
  B,
  Tau,
  Sigma_E,
  epi_pairs = NULL,
  epi_effects = NULL,
  seed = NULL
)
}
\description{
M2: 因果链模型
Y = (X\emph{B + eps) \%}\% solve(I - t(Tau))
协方差传播: Sigma_Y = (I-Tau)^{-1} \%\emph{\% Sigma_base \%}\% (I-t(Tau))^{-1}
}
